Clean up, prepare config for kopano-spamd setup, make learning on startup and deleting an option

+++ inotify-spamlearn.py
@@ -2,13 +2,11 @@
 # -*- coding: utf-8 -*-
 # vim: ai ts=4 sts=4 et sw=4
 
-
+import inotify.adapters
 import configparser
 import logging
 import os
 import subprocess
-
-import inotify.adapters
 
 
 def getconfig():
@@ -17,44 +15,68 @@ def getconfig():
         config.read('inotify-spamlearn.cfg')
         path = config.get('path', 'watch')
         learncmd = config.get('spam', 'learncmd')
-        return (path, learncmd)
+        delete = config.getboolean('spam', 'delete')
+        startup = config.getboolean('spam', 'startup')
+        logfile = config.get('logs', 'logfile')
+        return (path, learncmd, logfile, delete, startup)
     except:
         exit('Configuration error, please check \'inotify-spamlearn.cfg\'')
 
 
-def learn(filename, spamcmd):
+def learn(filename, spamcmd, delete):
     cmd = ' '.join([spamcmd, filename])
+    if os.path.exists(filename):
+        try:
+            p = subprocess.Popen(cmd.split(' '), stdout=subprocess.PIPE)
+            learning, output_err = p.communicate()
+            logging.info('Learn [%s]: %s' % (filename, str(learning.decode('utf-8')).strip('\n')))
 
+        except Exception as e:
+            logging.warning('Learn failed [{}]: '.format(e))
+
+        else:
+            if delete:
+                logging.info('Removing file: {}'.format(filename))
+                os.remove(filename)
+    else:
+        logging.warning('File %s does not exist (anymore)' % filename)
+
+
+def start():
+    (path, learncmd, logfile, delete, startup) = getconfig()
+    logging.basicConfig(filename=logfile, format='%(asctime)s %(message)s', level=logging.INFO)
+    logging.info('Starting Process')
+
+    if startup:
+        try:
+            logging.info('Looking for existing spam')
+            for spam in os.listdir(path):
+                learn('/'.join([path, spam]), learncmd, delete)
+        except:
+            logging.error('Cannot open path {}'.format(path))
+
+    i = inotify.adapters.Inotify()
     try:
-        p = subprocess.Popen(cmd.split(' '), stdout=subprocess.PIPE)
-        learning, output_err = p.communicate()
-        logging.info('Learn [%s]: %s' % (filename, str(learning.decode('utf-8')).strip('\n')))
-
-    except Exception as e:
-        logging.warning('Learn failed [%s]: ' % e)
+        i.add_watch(path)
+    except Exception:
+        logging.error('Cannot start inotify watch for {}'.format(path))
 
     else:
-        logging.info('Removing file: %s' % filename)
-        os.remove(filename)
-
-
-def main():
-    (path, learncmd) = getconfig()
-    logging.basicConfig(filename='example.log', format='%(asctime)s %(message)s', level=logging.INFO)
-    logging.info('Starting Inotify Learning')
-    i = inotify.adapters.Inotify()
-    i.add_watch(path)
+        logging.info('Inotify Learning Started')
 
     try:
         for event in i.event_gen():
             if event is not None:
                 (header, type_names, watch_path, filename) = event
-                if type_names[0] == 'IN_CLOSE_WRITE' and filename.split('.')[1] == 'eml':
-                    print('file written %s' % filename)
-                    learn('/'.join([path, filename]), learncmd)
+                if type_names[0] == 'IN_CLOSE_WRITE':
+                    learn('/'.join([path, filename]), learncmd, delete)
 
     finally:
         i.remove_watch(path)
+
+
+def main():
+    start()
 
 
 if __name__ == '__main__':
